Minor final changes

+++ MuyGPyS/examples/fast_posterior_mean.py
@@ -29,7 +29,7 @@ from MuyGPyS.examples.regress import _decide_and_make_regressor
 from MuyGPyS.gp import MuyGPS, MultivariateMuyGPS as MMuyGPS
 from MuyGPyS.gp.tensors import fast_nn_update
 from MuyGPyS.neighbors import NN_Wrapper
-from MuyGPyS._src.gp.tensors import _pairwise_tensor as pairwise_tensor
+from MuyGPyS.gp.tensors import pairwise_tensor
 
 
 def make_fast_regressor(
@@ -70,7 +70,7 @@ def make_fast_regressor(
     nn_indices, _ = nbrs_lookup.get_batch_nns(
         mm.arange(0, num_training_samples)
     )
-    nn_indices = nn_indices.astype(int)
+    nn_indices = nn_indices
     nn_indices = fast_nn_update(nn_indices)
 
     train_nn_targets = train_targets[nn_indices]
@@ -120,7 +120,7 @@ def make_fast_multivariate_regressor(
     nn_indices, _ = nbrs_lookup.get_batch_nns(
         mm.arange(0, num_training_samples)
     )
-    nn_indices = nn_indices.astype(int)
+    nn_indices = nn_indices
 
     nn_indices = fast_nn_update(nn_indices)
     pairwise_diffs_fast = pairwise_tensor(train_features, nn_indices)
